Refactor: better name for class handling quota
