+++ src/main/kotlin/no/nav/arbeidsplassen/emailer/sendmail/ScheduledOutboxEmailSender.kt
@@ -1,12 +1,12 @@
 package no.nav.arbeidsplassen.emailer.sendmail
 
 import net.javacrumbs.shedlock.spring.annotation.SchedulerLock
-import no.nav.arbeidsplassen.emailer.sendmail.LimitHandler.Companion.PENDING_EMAIL_CRON
-import no.nav.arbeidsplassen.emailer.sendmail.LimitHandler.Companion.PENDING_EMAIL_LOCK_AT_LEAST_FOR
-import no.nav.arbeidsplassen.emailer.sendmail.LimitHandler.Companion.PENDING_EMAIL_LOCK_AT_MOST_FOR
-import no.nav.arbeidsplassen.emailer.sendmail.LimitHandler.Companion.RETRY_EMAIL_CRON
-import no.nav.arbeidsplassen.emailer.sendmail.LimitHandler.Companion.RETRY_EMAIL_LOCK_AT_LEAST_FOR
-import no.nav.arbeidsplassen.emailer.sendmail.LimitHandler.Companion.RETRY_EMAIL_LOCK_AT_MOST_FOR
+import no.nav.arbeidsplassen.emailer.sendmail.EmailQuota.Companion.PENDING_EMAIL_CRON
+import no.nav.arbeidsplassen.emailer.sendmail.EmailQuota.Companion.PENDING_EMAIL_LOCK_AT_LEAST_FOR
+import no.nav.arbeidsplassen.emailer.sendmail.EmailQuota.Companion.PENDING_EMAIL_LOCK_AT_MOST_FOR
+import no.nav.arbeidsplassen.emailer.sendmail.EmailQuota.Companion.RETRY_EMAIL_CRON
+import no.nav.arbeidsplassen.emailer.sendmail.EmailQuota.Companion.RETRY_EMAIL_LOCK_AT_LEAST_FOR
+import no.nav.arbeidsplassen.emailer.sendmail.EmailQuota.Companion.RETRY_EMAIL_LOCK_AT_MOST_FOR
 import org.slf4j.LoggerFactory
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Service
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Service
 @Service
 class ScheduledOutboxEmailSender(
     private val emailService: EmailService,
-    private val limitHandler: LimitHandler,
+    private val emailQuota: EmailQuota,
     private val outboxEmailRepository: OutboxEmailRepository
 ) {
     companion object {
@@ -28,7 +28,7 @@ class ScheduledOutboxEmailSender(
         lockAtMostFor = PENDING_EMAIL_LOCK_AT_MOST_FOR
     )
     fun sendPendingEmails() {
-        val numberOfEmailsToSend = limitHandler.emailsToSend()
+        val numberOfEmailsToSend = emailQuota.emailsToSend()
 
         if (numberOfEmailsToSend == 0) {
             LOG.debug("No quota left for sending emails")
@@ -49,7 +49,7 @@ class ScheduledOutboxEmailSender(
     @Scheduled(cron = RETRY_EMAIL_CRON)
     @SchedulerLock(name = "retryFailedEmails", lockAtLeastFor = RETRY_EMAIL_LOCK_AT_LEAST_FOR, lockAtMostFor = RETRY_EMAIL_LOCK_AT_MOST_FOR)
     fun retryFailedEmails() {
-        val numberOfEmailsToSend = limitHandler.emailsToRetry()
+        val numberOfEmailsToSend = emailQuota.emailsToRetry()
 
         if (numberOfEmailsToSend == 0) {
             LOG.debug("No quota left for retrying failed emails")
